Organize modules into directories, and move tests accordingly
