+++ packages/ui-common/__tests__/components/MultiAgentAccelerator/ThoughtBubbles/ThoughtBubbleEdge.test.tsx
@@ -1,8 +1,8 @@
 import {render} from "@testing-library/react"
 import {Position} from "@xyflow/react"
 
-import {withStrictMocks} from "../../../../../__tests__/common/strictMocks"
-import {ThoughtBubbleEdge} from "../../../components/MultiAgentAccelerator/ThoughtBubbleEdge"
+import {withStrictMocks} from "../../../../../../__tests__/common/strictMocks"
+import {ThoughtBubbleEdge} from "../../../../components/MultiAgentAccelerator/ThoughtBubbles/ThoughtBubbleEdge"
 
 describe("ThoughtBubbleEdge", () => {
     withStrictMocks()
