+++ packages/ui-common/index.ts
@@ -40,7 +40,7 @@ export * from "./components/Common/notification"
 export * from "./components/ErrorPage/ErrorBoundary"
 export * from "./components/MultiAgentAccelerator/AgentConversations"
 export * from "./components/MultiAgentAccelerator/AgentCounts"
-export * from "./components/MultiAgentAccelerator/AgentFlow"
+export * from "./components/MultiAgentAccelerator/AgentFlow/AgentFlow"
 export * from "./components/MultiAgentAccelerator/MultiAgentAccelerator"
 export * from "./components/MultiAgentAccelerator/Sidebar/Sidebar"
 export * from "./components/MultiAgentAccelerator/TemporaryNetworks"
